+++ obsinfo/instrument_components/__init__.py
@@ -1,2 +1,0 @@
-from .instrument_components import instrument_components
-from .instrument_component import instrument_component