+++ obsinfo/instrumentation/__init__.py
@@ -1,2 +1,3 @@
-from .instrumentation import instrumentation
-from .instrument import instrument
+from .instrumentation import Instrumentation
+from .instrumentation import Instrument
+from .instrument_components import Instrument_components